+++ Src/Web/Web.Nuget.Tests/ConfigurationHelpers.cs
@@ -2,23 +2,33 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Diagnostics;
     using System.IO;
-    using System.Reflection;
     using System.Xml;
     using System.Xml.Linq;
     using Microsoft.Web.XmlTransform;
 
     public static class ConfigurationHelpers
     {
-        private const string ApplicationInsightsConfigInstall = "Microsoft.ApplicationInsights.Resources.ApplicationInsights.config.install.xdt";
-        private const string ApplicationInsightsConfigUninstall = "Microsoft.ApplicationInsights.Resources.ApplicationInsights.config.uninstall.xdt";
+        private const string ApplicationInsightsConfigInstallNet40 = "Microsoft.ApplicationInsights.Resources.net40.ApplicationInsights.config.install.xdt";
+        private const string ApplicationInsightsConfigInstallNet45 = "Microsoft.ApplicationInsights.Resources.net45.ApplicationInsights.config.install.xdt";
+        private const string ApplicationInsightsConfigUninstallNet40 = "Microsoft.ApplicationInsights.Resources.net40.ApplicationInsights.config.uninstall.xdt";
+        private const string ApplicationInsightsConfigUninstallNet45 = "Microsoft.ApplicationInsights.Resources.net45.ApplicationInsights.config.uninstall.xdt";
         private const string ApplicationInsightsTransform = "Microsoft.ApplicationInsights.Resources.ApplicationInsights.config.transform";
         
         private static readonly XNamespace XmlNamespace = "http://schemas.microsoft.com/ApplicationInsights/2013/Settings";
+        private static string applicationInsightsConfigInstall = ApplicationInsightsConfigInstallNet40;
+        private static string applicationInsightsConfigUninstall = ApplicationInsightsConfigUninstallNet40;
+
+        public static void ConfigureNet45()
+        {
+            applicationInsightsConfigInstall = ApplicationInsightsConfigInstallNet45;
+            applicationInsightsConfigUninstall = ApplicationInsightsConfigUninstallNet45;
+        }
 
         public static string GetEmptyConfig()
         {
-            Stream stream = typeof(TelemetryInitailizersTests).Assembly.GetManifestResourceStream(ApplicationInsightsTransform);
+            Stream stream = typeof(TelemetryInitailizersTestsNet40).Assembly.GetManifestResourceStream(ApplicationInsightsTransform);
             using (var reader = new StreamReader(stream))
             {
                 return reader.ReadToEnd();
@@ -47,19 +57,20 @@
 
         public static XDocument InstallTransform(string sourceXml)
         {
-            return Transform(sourceXml, ApplicationInsightsConfigInstall);
+            Debug.WriteLine(applicationInsightsConfigInstall);
+            return Transform(sourceXml, applicationInsightsConfigInstall);
         }
 
         public static XDocument UninstallTransform(string sourceXml)
         {
-            return Transform(sourceXml, ApplicationInsightsConfigUninstall);
+            return Transform(sourceXml, applicationInsightsConfigUninstall);
         }
 
         private static XDocument Transform(string sourceXml, string transformationFileResourceName)
         {
             using (var document = new XmlTransformableDocument())
             {
-                Stream stream = typeof(TelemetryInitailizersTests).Assembly.GetManifestResourceStream(transformationFileResourceName);
+                Stream stream = typeof(TelemetryInitailizersTestsNet40).Assembly.GetManifestResourceStream(transformationFileResourceName);
                 using (var reader = new StreamReader(stream))
                 {
                     string transform = reader.ReadToEnd();