+++ Src/Web/Web.Shared.Net.Tests/ApplicationInsightsHttpModuleTests.cs
@@ -36,7 +36,6 @@
         {
             ((IHttpModule)this.module.Target).Dispose();
             ((IHttpModule)this.module2.Target).Dispose();
-            Common.ActivityHelpers.StopRequestActivity();
         }
 
         [TestMethod]