+++ Src/Web/Web.Shared.Net.Tests/ClientIpHeaderTelemetryModuleTest.cs
@@ -4,6 +4,7 @@
     using System.Diagnostics;
     using System.Reflection;
     using System.Web;
+    using Microsoft.ApplicationInsights.Common;
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Web.Helpers;
     using Microsoft.ApplicationInsights.Web.Implementation;
@@ -12,16 +13,23 @@
     [TestClass]
     public class ClientIpHeaderTelemetryInitializerTest
     {
+        [TestCleanup]
+        public void Cleanup()
+        {
+#if NET45
+            while (Activity.Current != null)
+            {
+                Activity.Current.Stop();
+            }
+#else
+            ActivityHelpers.CleanOperationContext();
+#endif
+        }
+
         [TestInitialize]
         public void TestInitialize()
         {
             Trace.WriteLine(Assembly.GetExecutingAssembly().FullName);
-        }
-
-        [TestCleanup]
-        public void Cleanup()
-        {
-            Common.ActivityHelpers.StopRequestActivity();
         }
 
         [TestMethod]