+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 - Report status code for the dependencies failed with non-protocol issue like DNS resolution or SSL shakeup problems.
 - Implemented automatic telemetry correlation: all telemetry reported within the scope of the request is correlated to RequestTelemetry reported for the request.
 - Implemented [Correlation HTTP protocol](https://github.com/lmolkova/correlation/blob/master/http_protocol_proposal_v1.md): default headers to pass Operation Root Id and Parent Id were changed. This version is backward compatible with previously supported headers. 
+- Dependency to System.Diagnostics.DiagnosticsSource package is added for Web SDK on .NET 4.5.
 - Improvements to exception statistics, e.g. 2 of each type of exception will be output via TrackException
+- New ```AspNetDiagnosticTelemetryModule``` introduced for Web SDK on .NET 4.5, it consumes events from [Microsoft.AspNet.TelemetryCorrelation package](https://github.com/aspnet/AspNetCorrelationIdTracker) about incoming Http requests.
+- Dependency to Microsoft.AspNet.TelemetryCorrelation package is added for Web SDK on .NET 4.5.
 
 ## Version 2.3.0
 - Includes all changes since 2.2.0 stable release.